+++ tests/math/test_ecdf.py
@@ -1,0 +1,23 @@
+import numpy as np
+import pytest
+from hypothesis import given
+from hypothesis import strategies as st
+from hypothesis.extra.pandas import series
+
+
+@given(s=series(dtype=np.number))
+def test_ecdf(s):
+    """A simple execution test."""
+    if s.isnull().sum() > 0:
+        with pytest.raises(ValueError):
+            x, y = s.ecdf()
+    else:
+        x, y = s.ecdf()
+        assert len(x) == len(y)
+
+
+@given(s=series(dtype=str))
+def test_ecdf_string(s):
+    """Test that type enforcement is in place."""
+    with pytest.raises(TypeError):
+        x, y = s.ecdf()